IMN-577 Add authorization mocks (#561)
