+++ packages/authorization-updater/test/utils.ts
@@ -3,13 +3,7 @@ import {
   writeInReadmodel,
 } from "pagopa-interop-commons-test";
 import { inject, afterEach } from "vitest";
-import {
-  Client,
-  EService,
-  clientKind,
-  generateId,
-  toReadModelEService,
-} from "pagopa-interop-models";
+import { Client, EService, toReadModelEService } from "pagopa-interop-models";
 import { readModelServiceBuilder } from "../src/readModelService.js";
 
 export const { cleanup, readModelRepository } = setupTestContainersVitest(
@@ -30,15 +24,3 @@ export const addOneEService = async (eservice: EService): Promise<void> => {
 export const addOneClient = async (client: Client): Promise<void> => {
   await writeInReadmodel(client, clients);
 };
-
-export const getMockClient = (): Client => ({
-  id: generateId(),
-  consumerId: generateId(),
-  name: "Test client",
-  purposes: [],
-  description: "Client description",
-  users: [],
-  kind: clientKind.consumer,
-  createdAt: new Date(),
-  keys: [],
-});
