+++ packages/commons-test/src/testUtils.ts
@@ -26,6 +26,10 @@ import {
   purposeVersionState,
   Document,
   AgreementAttribute,
+  Key,
+  Client,
+  clientKind,
+  keyUse,
 } from "pagopa-interop-models";
 import { AuthData } from "pagopa-interop-commons";
 
@@ -222,4 +226,25 @@ export const getMockDocument = (): Document => ({
   contentType: "json",
   checksum: "checksum",
   uploadDate: new Date(),
+});
+
+export const getMockClient = (): Client => ({
+  id: generateId(),
+  consumerId: generateId(),
+  name: "Test client",
+  purposes: [],
+  description: "Client description",
+  users: [],
+  kind: clientKind.consumer,
+  createdAt: new Date(),
+  keys: [],
+});
+
+export const getMockKey = (): Key => ({
+  name: "test key",
+  createdAt: new Date(),
+  kid: generateId(),
+  encodedPem: generateId(),
+  algorithm: "",
+  use: keyUse.sig,
 });
